refactor after first PR: searchBar component changed and renamed to hold all filter related functions, the DayPicker and related logic
