+++ pages/emotion-records.js
@@ -1,38 +1,16 @@
 import EmotionRecordsList from "@/components/EmotionRecordsList";
 import {
   StyledFlexColumnWrapper,
-  StyledButton,
   StyledStandardLink,
 } from "@/SharedStyledComponents";
 import styled from "styled-components";
-import SearchBar from "@/components/SearchBar";
-import { useState, useEffect } from "react";
-import Fuse from "fuse.js";
-import { DayPicker } from "react-day-picker";
+import FilterEmotionEntries from "@/components/FilterEmotionEntries";
+import { useState, useCallback } from "react";
+import HeartOutlineIcon from "../public/heart-outline.svg";
 import CalendarIcon from "/public/calendar.svg";
 
-const StyledCalendarIcon = styled(CalendarIcon)`
-  width: 1.5rem;
-  display: inline;
-  vertical-align: bottom;
-`;
-
-const StyledFilterSection = styled.section`
-  width: 100%;
-  background-color: var(--main-bright);
-  padding: 0.5rem;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  justify-content: center;
-  align-items: center;
-  position: fixed;
-  top: 100px;
-  z-index: 1;
-`;
-
 const StyledTextMessage = styled.p`
-  margin-top: 150px;
+  margin-top: 4rem;
   text-align: center;
   line-height: 3;
 `;
@@ -42,63 +20,22 @@ const StyledLink = styled(StyledStandardLink)`
   background-color: var(--button-background);
 `;
 
-const StyledFilterButton = styled(StyledButton)`
-  background-color: ${({ $active }) =>
-    $active ? "var(--button-background)" : "var(--main-bright)"};
-  margin: 0;
-  padding: 0.2rem;
-  width: auto;
+const StyledHeartSymbol = styled(HeartOutlineIcon)`
+  width: 1.4rem;
+  display: inline;
+  position: relative;
+  top: 5px;
 `;
 
-const StyledEmotionRecordsTitle = styled.h1`
-  font-weight: 600;
+const StyledCalendarIcon = styled(CalendarIcon)`
+  width: 1.5rem;
+  display: inline;
+  vertical-align: bottom;
 `;
 
-const StyledButtonGroup = styled.div`
-  display: flex;
-  justify-content: center;
-  flex-wrap: wrap;
-  padding: 0.6rem;
-  width: 90vw;
-  gap: 10px;
-`;
-
-// styles for DayPicker
-const StyledBackground = styled.div`
-  background-color: black;
-  opacity: 0.5;
-  position: fixed;
-  inset: 0;
-  z-index: 2;
-`;
-
-const WrapperForDayPicker = styled.section`
-  position: fixed;
-  right: calc(50% - 170px);
-  top: 9rem;
-  width: 340px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  z-index: 4;
-  background-color: var(--main-bright);
-  border-radius: 6px;
-  padding: 1rem;
-`;
-
-const WrapperForNavigation = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-evenly;
-  padding: 0.5rem;
-`;
-
-const StyledNavButton = styled(StyledButton)`
-  width: 5rem;
-  padding: 0.2rem;
-  margin: 0.5rem;
-  opacity: ${(props) => (props.disabled ? "0.25" : "1")};
+const StyledDateIndicator = styled.p`
+  text-align: center;
+  margin: 2rem auto 1rem;
 `;
 
 const StyledParagraph = styled.p`
@@ -114,6 +51,7 @@ export default function EmotionRecords({
   const [searchTerm, setSearchTerm] = useState();
   const [filteredEntries, setFilteredEntries] = useState(emotionEntries);
   const [shownEntries, setShownEntries] = useState(emotionEntries);
+  const [selectedTime, setSelectedTime] = useState();
   const [buttonState, setButtonState] = useState({
     todayButton: true,
     name: "todayButton",
@@ -122,141 +60,23 @@ export default function EmotionRecords({
     daysAgo: 0,
   });
 
-  //useStates for DayPicker
-  const [month, setMonth] = useState(new Date());
-  const [selectedTime, setSelectedTime] = useState();
-  const [showDayPicker, setShowDayPicker] = useState(false);
-  //
+  // handler-functions used in a useEffect after passed to FilterEmotionEntries are wrapped into useCallback hook here
 
-  const filterButtons = [
-    { name: "allButton", label: "All" },
-    { name: "todayButton", label: "Today", singleComparison: true, daysAgo: 0 },
-    {
-      name: "yesterdayButton",
-      label: "Yesterday",
-      singleComparison: true,
-      daysAgo: 1,
-    },
-    { name: "weekButton", label: " Last Week", daysAgo: 7 },
-    { name: "monthButton", label: "Last Month", daysAgo: 30 },
-    { name: "highlightedButton", label: "Highlighted" },
-    {
-      name: "datePicker",
-      label: `Custom`,
-      icon: <StyledCalendarIcon />,
-      setShow: true,
-    },
-  ];
+  const handleSetFilterEntries = useCallback((filteredObject) => {
+    setFilteredEntries(filteredObject);
+  }, []);
 
-  // sets filteredEntries depending on buttonState; reacts to changes of emotionEntrie in case of e.g. deletion
+  const handleSetShownEntries = useCallback((entriesObjectAfterSearch) => {
+    setShownEntries(entriesObjectAfterSearch);
+  }, []);
 
-  useEffect(() => {
-    // helper functions
-    function getDate(daysToGoBack) {
-      const currentTime = new Date();
-      const date = new Date(
-        new Date(
-          currentTime.setDate(currentTime.getDate() - daysToGoBack)
-        ).toDateString()
-      ).getTime();
+  function handleSetButtonState(buttonObject) {
+    setButtonState(buttonObject);
+  }
 
-      return date;
-    }
-
-    function getFilteredEntries(boolean, firstDate, secondDate) {
-      const filteredEntries = emotionEntries.filter((entry) => {
-        const entrieDate = new Date(
-          new Date(entry.isoDate).toDateString()
-        ).getTime();
-
-        const result = boolean
-          ? entrieDate === firstDate
-          : entrieDate >= firstDate && entrieDate <= secondDate;
-        return result;
-      });
-      return filteredEntries;
-    }
-
-    const { daysAgo, singleComparison } = buttonState;
-    //
-
-    setFilteredEntries(() => {
-      if (
-        buttonState.todayButton ||
-        buttonState.yesterdayButton ||
-        buttonState.weekButton ||
-        buttonState.monthButton
-      ) {
-        const todayOrEarlier = getDate(daysAgo);
-        const today = getDate(0);
-
-        const buttonFilteredEntries = getFilteredEntries(
-          singleComparison,
-          todayOrEarlier,
-          today
-        );
-        return buttonFilteredEntries;
-      }
-      if (buttonState.highlightedButton) {
-        const highlightedEntries = emotionEntries.filter(
-          (entry) => entry.isHighlighted
-        );
-        return highlightedEntries;
-      }
-      if (buttonState.datePicker) {
-        if (!selectedTime) {
-          return emotionEntries;
-        } else {
-          const selectedStartDate = new Date(
-            selectedTime.from.toDateString()
-          ).getTime();
-
-          const selectedEndDate =
-            selectedTime.to &&
-            new Date(selectedTime.to.toDateString()).getTime();
-
-          const secondDateNotSelected = !selectedEndDate;
-
-          const customFilteredEntries = getFilteredEntries(
-            secondDateNotSelected,
-            selectedStartDate,
-            selectedEndDate
-          );
-
-          return customFilteredEntries;
-        }
-      } else {
-        return emotionEntries;
-      }
-    });
-  }, [buttonState, emotionEntries, selectedTime]);
-
-  // searches only filteredEntries and sets shownEntries accordingly
-  useEffect(() => {
-    if (!searchTerm) {
-      setShownEntries(filteredEntries);
-      return;
-    }
-
-    const fuse = new Fuse(filteredEntries, {
-      includeScore: true,
-      threshold: 0.4,
-      keys: [
-        "timeAndDate",
-        "tensionLevel",
-        "trigger",
-        "intensity",
-        "notes",
-        "category",
-        "emotion",
-        "subemotion",
-      ],
-    });
-
-    const results = fuse.search(searchTerm);
-    const items = results.map((result) => result.item);
-    setShownEntries(items);
-  }, [searchTerm, filteredEntries]);
+  function handleSetSelectedTime(time) {
+    setSelectedTime(time);
+  }
 
   function handleSearch(input) {
     setSearchTerm(input);
@@ -272,100 +92,83 @@ export default function EmotionRecords({
     return date;
   }
 
+  function DisplayDate() {
+    return (
+      selectedTime && (
+        <StyledParagraph>
+          Your Selection:<br></br>
+          {getFormattedDate(selectedTime.from)}
+          {selectedTime.to &&
+            selectedTime.from.toString() !== selectedTime.to.toString() &&
+            " - " + getFormattedDate(selectedTime.to)}
+        </StyledParagraph>
+      )
+    );
+  }
+
   return (
-    <StyledFlexColumnWrapper>
-      <StyledFilterSection>
-        <StyledEmotionRecordsTitle>Recorded Emotions</StyledEmotionRecordsTitle>
-        <SearchBar onSearch={handleSearch} />
-        <StyledButtonGroup>
-          {filterButtons.map(
-            ({ name, label, singleComparison, daysAgo, setShow, icon }) => (
-              <StyledFilterButton
-                key={name}
-                $active={buttonState[name] && true}
-                onClick={() => {
-                  setButtonState({
-                    [name]: true,
-                    singleComparison,
-                    daysAgo,
-                  });
-                  setShow && setShowDayPicker(!showDayPicker);
-                }}
-              >
-                {label}
-                {icon}
-              </StyledFilterButton>
+    <>
+      <StyledFlexColumnWrapper>
+        <FilterEmotionEntries
+          emotionEntries={emotionEntries}
+          filteredEntries={filteredEntries}
+          buttonState={buttonState}
+          searchTerm={searchTerm}
+          selectedTime={selectedTime}
+          onSearch={handleSearch}
+          changeShownEntries={handleSetShownEntries}
+          changeButtonState={handleSetButtonState}
+          changeFilterEntries={handleSetFilterEntries}
+          changeSelectedTime={handleSetSelectedTime}
+          DisplayDate={DisplayDate}
+        />
+        {buttonState.datePicker && !searchTerm ? (
+          selectedTime ? (
+            <DisplayDate />
+          ) : (
+            <StyledDateIndicator>
+              Click the calendar <StyledCalendarIcon /> and select a date
+            </StyledDateIndicator>
+          )
+        ) : null}
+        {shownEntries.length === 0 &&
+          (filteredEntries.length === 0 ? (
+            buttonState.highlightedButton ? (
+              <StyledTextMessage>
+                You haven&apos;t highlighted any Entries yet. Click the{" "}
+                <StyledHeartSymbol /> on a Entry to highlight it.`
+              </StyledTextMessage>
+            ) : buttonState.todayButton ? (
+              <StyledTextMessage>
+                You haven&apos;t made any Entries today.<br></br>
+                <StyledLink href="./">add Entry &rarr;</StyledLink>
+              </StyledTextMessage>
+            ) : (
+              <StyledTextMessage>sorry, nothing found</StyledTextMessage>
             )
-          )}
-        </StyledButtonGroup>
-      </StyledFilterSection>
-      {emotionEntries.length !== 0 && (
-        <>
-          {showDayPicker && (
-            <>
-              <StyledBackground />
-              <WrapperForDayPicker>
-                <DayPicker
-                  mode="range"
-                  selected={selectedTime}
-                  onSelect={setSelectedTime}
-                  toDate={new Date()}
-                  month={month}
-                  onMonthChange={setMonth}
-                />
-                {selectedTime ? (
-                  <StyledParagraph>
-                    Your Selection:<br></br>
-                    {getFormattedDate(selectedTime.from)}
-                    {selectedTime.to &&
-                      selectedTime.from.toString() !==
-                        selectedTime.to.toString() &&
-                      " - " + getFormattedDate(selectedTime.to)}
-                  </StyledParagraph>
-                ) : (
-                  <StyledParagraph>
-                    Select a single date or a range of dates
-                  </StyledParagraph>
-                )}
-                <WrapperForNavigation>
-                  <StyledNavButton
-                    disabled={
-                      month.getMonth() === new Date().getMonth() ? true : false
-                    }
-                    onClick={() => setMonth(new Date())}
-                  >
-                    Today
-                  </StyledNavButton>
-                  <StyledNavButton
-                    disabled={selectedTime ? false : true}
-                    onClick={() => setSelectedTime()}
-                  >
-                    Reset
-                  </StyledNavButton>
-                  <StyledNavButton onClick={() => setShowDayPicker(false)}>
-                    Ok
-                  </StyledNavButton>
-                </WrapperForNavigation>
-              </WrapperForDayPicker>
-            </>
-          )}
-          <EmotionRecordsList
-            onDeleteEmotionEntry={onDeleteEmotionEntry}
-            toggleHighlight={toggleHighlight}
-            shownEntries={shownEntries}
-            filteredEntries={filteredEntries}
-            buttonState={buttonState}
-            searchTerm={searchTerm}
-            selectedTime={selectedTime}
-          />
-        </>
-      )}
-      {emotionEntries.length === 0 && (
-        <StyledTextMessage>
-          You haven&apos;t made any Entries yet.<br></br>
-          <StyledLink href="./">add Entry &rarr;</StyledLink>
-        </StyledTextMessage>
-      )}
-    </StyledFlexColumnWrapper>
+          ) : (
+            <StyledTextMessage>sorry, nothing found</StyledTextMessage>
+          ))}
+
+        {shownEntries && (
+          <>
+            <EmotionRecordsList
+              onDeleteEmotionEntry={onDeleteEmotionEntry}
+              toggleHighlight={toggleHighlight}
+              shownEntries={shownEntries}
+              filteredEntries={filteredEntries}
+              buttonState={buttonState}
+              searchTerm={searchTerm}
+            />
+          </>
+        )}
+        {emotionEntries.length === 0 && (
+          <StyledTextMessage>
+            You haven&apos;t made any Entries yet.<br></br>
+            <StyledLink href="./">add Entry &rarr;</StyledLink>
+          </StyledTextMessage>
+        )}
+      </StyledFlexColumnWrapper>
+    </>
   );
 }
